+++ bot/utils.py
@@ -11,6 +11,7 @@ from telegram import (
 )
 from telegram.ext import ConversationHandler, TypeHandler
 
+from bot.constants import REMOVE_KEYBOARD_KEY
 from bot.twitter import build_sticker
 from bot.userdata import CCT
 
@@ -50,8 +51,23 @@ def default_message(update: Update, _: CCT) -> None:
     )
 
 
-def conversation_timeout(update: Update, _: CCT) -> int:
-    """Informs the user that the operation has timed out and ends the conversation.
+def remove_reply_markup(context: CCT) -> None:
+    """Removes the reply markup of the message stored in ``context.chat_data[REMOVE_KEYBOARD]``,
+    if any.
+
+    Args:
+        context: The callback context as provided by the dispatcher.
+    """
+    if not context.chat_data:
+        return
+    message = context.chat_data.get(REMOVE_KEYBOARD_KEY, None)
+    if isinstance(message, Message):
+        message.edit_reply_markup(None)
+
+
+def conversation_timeout(update: Update, context: CCT) -> int:
+    """Informs the user that the operation has timed out, calls :meth:`remove_reply_markup` and
+    ends the conversation.
 
     Args:
         update: The Telegram update.
@@ -61,6 +77,7 @@ def conversation_timeout(update: Update, _: CCT) -> int:
         int: :attr:`telegram.ext.ConversationHandler.END`.
     """
     cast(User, update.effective_user).send_message('Operation timed out. Aborting.')
+    remove_reply_markup(context)
 
     return ConversationHandler.END
 
@@ -70,17 +87,19 @@ TIMEOUT_HANDLER = TypeHandler(Update, conversation_timeout)
 conversations."""
 
 
-def conversation_fallback(update: Update, _: CCT) -> int:
-    """Informs the user that the input was invalid and ends the conversation.
+def conversation_fallback(update: Update, context: CCT) -> int:
+    """Informs the user that the input was invalid, calls :meth:`remove_reply_markup` and
+    ends the conversation.
 
     Args:
         update: The Telegram update.
-        _: The callback context as provided by the dispatcher.
+        context: The callback context as provided by the dispatcher.
 
     Returns:
         int: The next state.
     """
     cast(User, update.effective_user).send_message('Invalid input. Aborting operation.')
+    remove_reply_markup(context)
 
     return ConversationHandler.END